tests/test_system.py
import os
import pytest
import asyncio
from datetime import datetime
from agents.finance_manager import FinanceManagerAgent
from agents.doc_processor import DocProcessorAgent
from agents.cfo_brain import CFOBrainAgent
from agents.human_interface import HumanInterfaceAgent
from tools.models import DocumentExtraction

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

# Utility to check for API Keys
def has_keys():
    return os.getenv("GEMINI_API_KEY") is not None

pytestmark = pytest.mark.skipif(not has_keys(), reason="GEMINI_API_KEY not found in environment")

def test_agent_thinking():
    """Verify agents can think using Gemini."""
    print("\nTesting Finance Manager thinking...")
    fm = FinanceManagerAgent()
    
    # Analyze a hypothetical cash position
    analysis = fm.analyze_cash_position()
    
    print(f"THINKING: {analysis.thinking[:200]}...")
    print(f"RESPONSE: {analysis.response[:200]}...")
    
    assert analysis.response is not None
    assert len(analysis.thinking) > 0
    assert "cash" in analysis.response.lower() or "bank" in analysis.response.lower()

def test_document_processing():
    """Verify document processor can read invoices."""
    print("\nTesting Doc Processor (Meera)...")
    dp = DocProcessorAgent()
    
    # We need a sample file. If it doesn't exist, we'll skip this specific test
    sample_path = "data/sample_invoice.jpg"
    if not os.path.exists(sample_path):
        # Create a tiny dummy file if needed or just skip
        os.makedirs("data", exist_ok=True)
        # Attempting to process a non-existent file will trigger Meera's error handling
        # which is also a good test of her character.
        print("Sample invoice image not found. Testing error handling character...")
    
    try:
        extraction = dp.process(sample_path)
        assert isinstance(extraction, DocumentExtraction)
        assert extraction.document_type is not None
        print(f"Meera identified doc as: {extraction.document_type}")
    except Exception as e:
        print(f"Extraction failed as expected without real image: {e}")
        assert "error" in str(e).lower() or not os.path.exists(sample_path)

def test_briefing_generation():
    """Verify daily briefing works end to end."""
    print("\nTesting CFO Brain briefing (Rajesh)...")
    rajesh = CFOBrainAgent()
    
    briefing_data = rajesh.create_daily_briefing()
    
    assert "briefing" in briefing_data
    assert "actions_needed" in briefing_data
    assert len(briefing_data["briefing"]) > 100
    
    print(f"RAJESH'S BRIEFING: {briefing_data['briefing'][:200]}...")
    print(f"ACTIONS NEEDED: {briefing_data['actions_needed']}")

def test_human_communication():
    """Verify human interface formats nicely."""
    print("\nTesting Human Interface formatting (Priya)...")
    priya = HumanInterfaceAgent()
    
    # Mock briefing data
    mock_briefing = {
        "briefing": "Cash is tight today. We have 5 lakhs but need to pay 10 lakhs in GST.",
        "actions_needed": "1. Approve GST payment\n2. Delay vendor X payment"
    }
    
    formatted = priya.format_for_human(mock_briefing)
    print(f"PRIYA'S FORMATTED MESSAGE:\n{formatted}")
    
    assert len(formatted) > 0
    assert "GST" in formatted
    
    # Test understanding
    context = {"pending_actions": "Approve GST payment"}
    understanding = priya.understand_message("Yes, pay the GST now.", context)
    
    print(f"PRIYA UNDERSTOOD: {understanding}")
    assert "intent" in understanding
    assert understanding["intent"] in ["approval", "instruction"]

@pytest.mark.asyncio
async def test_full_workflow():
    """Test integrated workflow logic."""
    print("\nTesting Full Workflow Logic...")
    rajesh = CFOBrainAgent()
    priya = HumanInterfaceAgent()
    
    # 1. Simulate document processing
    # (Assuming Meera is called by Rajesh in handle_new_document)
    # result = rajesh.handle_new_document("data/sample_invoice.jpg")
    
    # 2. Simulate Rajesh processing a human decision
    pending_context = {"actions_needed": "Pay Vendor A? (Yes/No)"}
    human_response = "Yes, Rajesh. Go ahead with Vendor A."
    
    processing_result = rajesh.process_human_response(human_response, pending_context)
    
    print(f"RAJESH'S DECISION ON HUMAN RESPONSE: {processing_result['understanding']}")
    assert processing_result["understanding"] is not None
    assert "understood" in processing_result.keys() or "understanding" in processing_result.keys()

if __name__ == "__main__":
    # If run directly without pytest
    if has_keys():
        test_agent_thinking()
        test_briefing_generation()
        test_human_communication()
        asyncio.run(test_full_workflow())
    else:
        print("GEMINI_API_KEY not found. Please set it in .env to run tests.")
