requirements.txt
google-generativeai
opik
python-telegram-bot
gspread
google-auth
pydantic
python-dotenv
Pillow
PyPDF2
